url, facebook, twitter, etc should be URL fields in Database, not Text - also Email fields
When we display info in the template, we want to know it's a URL and to just display it.

So change field type, and if ALISS sends data that is NOT a URL, we will let the importer deal with it!

closes https://github.com/OpenDataServices/iCAN-Scot-Catalogue/issues/4

Reset migrations as there is no real data yet anyway and we aren't sure about changing types.